requirements.txt
line-bot-sdk
Flask
beautifulsoup4
bs4
requests
